feat: implement session management with database persistence
